widget/tests/window_bug593307_centerscreen.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<window title="Mozilla Bug 593307"
  width="100" height="100"
  onload="onload();"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js" />

<body xmlns="http://www.w3.org/1999/xhtml" id="body">
</body>

<script class="testbody" type="application/javascript">
<![CDATA[

function onload()
{
  var SimpleTest = window.opener.SimpleTest;
  SimpleTest.ok(window.screenX >= 0, "centerscreen window should not start offscreen (X coordinate): " + window.screenX);
  SimpleTest.ok(window.screenY >= 0, "centerscreen window should not start offscreen (Y coordinate): " + window.screenY);
  window.opener.finished();
}
]]>
</script>

</window>
